SystemMessenger::post() - The invoke function will not be executed immediately even it is on UI thread
